requirements.txt
click==8.0.4
deap==1.3.3
dysts==0.1
matplotlib==3.7.2
numba==0.58.1
numpy==1.22.3
pandas==1.5.3
pysindy==1.7.5
derivative==0.6.0
PyYAML==6.0.1
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.13.0
symengine==0.11.0
sympy==1.10.1
tqdm==4.65.0
